React Component for the Task
I used react in this solution. this is the relevant component for hitting the api endpoint provided for the countries data and displaying the said table per instructions. however, i'm unable to integrate the map api as at the time of this upload. been facing network issues till this minute. Thank you